move utils in e2e/ to src/e2e-app/
