+++ e2e/components/slide-toggle-e2e.spec.ts
@@ -1,5 +1,5 @@
 import {browser, element, by, Key} from 'protractor';
-import {expectToExist} from '../util/index';
+import {expectToExist} from '../../src/e2e-app/test-util/index';
 
 
 describe('slide-toggle', () => {
