+++ e2e/components/stepper-e2e.spec.ts
@@ -1,7 +1,7 @@
 import {browser, by, element, ElementFinder, ExpectedConditions} from 'protractor';
 import {Key} from 'selenium-webdriver';
-import {pressKeys} from '../util/actions';
-import {expectFocusOn, expectToExist} from '../util/asserts';
+import {pressKeys} from '../../src/e2e-app/test-util/actions';
+import {expectFocusOn, expectToExist} from '../../src/e2e-app/test-util/asserts';
 
 describe('stepper', () => {
   beforeEach(async () => await browser.get('/stepper'));
